Updated CSS to number theorem environments
This doesn't touch problems, exercises, etc, but it should number
theorems, axioms, conjectures, and so on -- each with their own separate
counter. Did some work to make it easier to change our minds on this
later. Side note: I don't have the entire ximera server running locally,
so pushing changes like this is a little risky, but it is "just the
CSS," so maybe okay. Cleaned up the loop and the array as well to make
it more understandable.